compilation tool migrated from gfortran to NVHPC
将 ASTR 项目从 gfortran 编译迁移至 NVIDIA NVHPC 26.1 (`nvfortran`) 编译器过程中遇到的源代码问题及对应修复。

| 组件 | 版本/路径 |
|------|-----------|
| NVHPC SDK | 26.1 (`/opt/nvidia/hpc_sdk/Linux_x86_64/26.1`) |
| MPI | HPC-X OpenMPI (`hpcx/bin/mpif90`) |
| HDF5 | 1.14.6 (`/opt/hdf5-1.14.6`, 已用 NVHPC 编译) |
| 编译器 | `h5pfc` → `nvfortran` |

关键编译标志：`-Mpreprocess -Mextend -traceback -g -O3 -DHDF5 -Kieee`

---

**文件**: `src/strings.F90` 第 2659–2705 行

**问题**: `visible` 函数内定义了一个 256 元素的字符参数数组 `chars(0:255)`，其中包含 `'M-^\'` 这样的字符串字面量。nvfortran 将字符串内的 `\'` 序列误解析为转义的单引号，导致引号匹配错误。

**报错信息**:
```
NVFORTRAN-S-0034-Syntax error at or near identifier m (strings.F90: 2669)
NVFORTRAN-S-0023-Syntax error - unbalanced brackets (strings.F90: 2687)
NVFORTRAN-S-0038-Symbol, chars, has not been explicitly declared
```

**根因分析**: 经过逐元素二分隔离测试，确认触发条件是 `'M-^\'` 这个特定的字符串——反斜杠 `\` 后紧跟单引号 `'` 被 nvfortran 解析器识别为类 C 语言的转义序列 `\'`（转义单引号），导致后续所有字符解析错乱。

`visible` 函数并未被项目中任何其他模块调用（仅 `strings` 模块的 `split` 被 `utility.F90` 使用），因此直接删除该函数是安全的。

**修复**:
1. 删除整个 `visible` 函数（第 2626–2705 行，含注释块）
2. 移除对应的 `PUBLIC visible` 声明（原第 52 行）

---

**文件**: `src/vtkio.F90` 第 38, 190, 358, 535 行

**问题**: NVHPC 不支持 `integer(kind=16)`（16 字节 / 128 位整数）。该变量 `ioff` 用作文件偏移量，64 位整数已足够。

**报错信息**:
```
NVFORTRAN-S-0081-Illegal selector - KIND parameter has unknown value for data type
```

**修复**: 模块级引入 `int64`，4 处局部声明改用标准可移植类型：
```diff
 module WriteVTK
+  use iso_fortran_env, only: int64
   implicit none
   ...
-     integer(kind=16) :: ioff
+     integer(int64) :: ioff
```

> `int64` (KIND=8) 上限 9.2 EB，对文件偏移量完全够用。

---

**文件**: `src/fludyna.F90`, `src/geom.F90`, `src/solver.F90`, `src/pp.F90`

**问题**: `isnan()` 是 gfortran 的内置扩展函数，不是标准 Fortran。NVHPC 不提供该函数。标准 Fortran 2008 的等价函数是 `ieee_is_nan()`，需通过 `use ieee_arithmetic` 引入。

**报错信息**:
```
NVFORTRAN-S-0038-Symbol, isnan, has not been explicitly declared
```

**涉及子程序** (共 8 个):

| 文件 | 子程序 | 行号范围 |
|------|--------|---------|
| `src/fludyna.F90` | `updateq` | 251–293 |
| `src/geom.F90` | `gridgeom` | 99–956 |
| `src/geom.F90` | `pointintriangle_tri` | 4143–4246 |
| `src/geom.F90` | `pointintriangle_nodes` | 4248–4321 |
| `src/solver.F90` | `src_tbl` | 401–487 |
| `src/pp.F90` | `solidgen_sphere_tri` | 1353–1577 |
| `src/pp.F90` | `solidgen_sphere` | 1590–1727 |
| `src/pp.F90` | `solidgen_circle` | 1739–1872 |

**修复** (分两步):

**步骤 A** — 全局替换函数名:
```bash
sed -i 's/isnan(/ieee_is_nan(/g' src/fludyna.F90 src/geom.F90 src/solver.F90 src/pp.F90
```

**步骤 B** — 在每个子程序的 `use` 声明块中添加:
```fortran
use ieee_arithmetic, only: ieee_is_nan
```

---

**文件**: `src/strings.F90` 第 5373 行

**问题**: NVHPC 不支持四精度浮点数 `real128`。`print_generic` 子程序中的 `SELECT TYPE` 分支引用了 `real(kind=real128)`，导致编译错误。

**报错信息**:
```
NVFORTRAN-S-0081-Illegal selector - KIND value must be non-negative
NVFORTRAN-S-0155-Duplicate TYPE IS real
```

**修复**: 用预处理器条件包裹，仅在非 NVHPC 编译器下编译该分支:
```diff
+ #ifndef __NVCOMPILER
      type is (real(kind=real128));     write(line(istart:),'(1pg0)') generic
+ #endif
```

> 注：NVHPC 预处理器同时定义 `__NVCOMPILER` 和 `__PGI` 宏。

---

**文件**: `build_nvhpc.sh`

**问题**: `h5pfc` 默认只链接 `-lhdf5_fortran -lhdf5`，但 ASTR 使用了 HDF5 Light (`h5lt`) API（`use h5lt`），需要额外链接 `-lhdf5_hl_fortran -lhdf5_hl`。

**报错信息**:
```
undefined reference to `h5lt_h5ltread_dataset_integer_kind_4_rank_1_'
undefined reference to `h5lt_h5ltmake_dataset_real_kind_8_rank_1_'
... (等 13 个未定义引用)
```

**修复**:
```bash
LIBS="-lhdf5_hl_fortran -lhdf5_hl -lz -lm"
```

---

| gfortran 标志 | nvfortran 标志 | 用途 |
|---------------|----------------|------|
| `-cpp` | `-Mpreprocess` | C 风格预处理 |
| `-ffree-line-length-none` | `-Mextend` | 132 列源代码行 |
| `-fbacktrace` | `-traceback` | 运行时错误回溯 |
| `-J <dir>` | `-module <dir>` | 模块文件输出目录 |
| `-fallow-argument-mismatch` | 不需要 | nvfortran 默认允许 |

---

```bash
bash build_nvhpc.sh

mpirun -np 8 ./bin/astr run
```

可执行文件: `bin/astr` (8.5 MB, ELF 64-bit, 动态链接, 带调试符号)

+++ src/vtkio.F90
@@ -3,6 +3,7 @@
 !+---------------------------------------------------------------------+
 module WriteVTK
   !
+  use iso_fortran_env, only: int64
   implicit none
   !
   Interface writeprvbin
@@ -34,7 +35,7 @@ module WriteVTK
     !
     integer :: int32,nbrvar,jvar
     integer :: int_bytesize,real_bytesize,nbt_scal,nbt_vec,nbt_xyz,nbt_ien,nbt_offset,nbt_etype
-    integer(16) :: ioff
+    integer(int64) :: ioff
     ! ip : le point actuel
     !
     real(4) :: float32
@@ -186,7 +187,7 @@ module WriteVTK
     !
     integer :: int32,nbrvar,jvar
     integer :: int_bytesize,real_bytesize,nbt_scal,nbt_vec!,nbt_xyz,nbt_ien,nbt_offset,nbt_etype
-    integer(16) :: ioff
+    integer(kind=8) :: ioff
     ! ip : le point actuel
     !
     real(4) :: float32
@@ -354,7 +355,7 @@ module WriteVTK
     !
     integer :: int32,unitf,nbrvar,n,jvar
     integer :: int_bytesize,real_bytesize,nbt_scal,nbt_vec,nbt_xyz,nbt_ien,nbt_offset,nbt_etype
-    integer(16) :: ioff
+    integer(8) :: ioff
     ! ip : le point actuel
     !
     real(4) :: float32
@@ -531,7 +532,7 @@ module WriteVTK
     !
     integer :: int32,nbrvar,jvar,kmax
     integer :: int_bytesize,real_bytesize,nbt_scal,nbt_vec!,nbt_xyz,nbt_ien,nbt_offset,nbt_etype
-    integer(16) :: ioff
+    integer(8) :: ioff
     ! ip : le point actuel
     !
     real(4) :: float32
